+++ web/src/lib/utils/api/getUser.ts
@@ -1,6 +1,6 @@
-import { API_BASE_URL } from '@/lib/utils/constants';
+import { API_BASE_URL } from '@/lib/constants/constants';
 
-import { User } from '@/lib/typing';
+import { User } from '@/lib/types';
 
 export default async function getUser(): Promise<User | null> {
   const response = await fetch(`${API_BASE_URL}/user`, {
